a Car Models
This update would add another function to the car brand model to get all
brands to display in the view. the controllers and view are edited to
implement this but are not yet finished the other controllers have been
worked on and the car_m_model.php ha had the class renamed to match the
convention.