Changes the brand view, delete and update buttons
Change the car_list view to only display the brand name on the left,
pressing this now will display the brand info in a card on the right.
Add a function in the model to support the new view type.
Change the controller to call thenew method and pass it to the view.